added tools registry to separate it out and avoid circular dependency
