+++ sim/tools/http/request.ts
@@ -1,5 +1,29 @@
 import { HttpMethod, TableRow, ToolConfig, ToolResponse } from '../types'
-import { transformTable } from '../utils'
+
+/**
+ * Transforms a table from the store format to a key-value object
+ * Local copy of the function to break circular dependencies
+ * @param table Array of table rows from the store
+ * @returns Record of key-value pairs
+ */
+const transformTable = (table: TableRow[] | null): Record<string, any> => {
+  if (!table) return {}
+
+  return table.reduce(
+    (acc, row) => {
+      if (row.cells?.Key && row.cells?.Value !== undefined) {
+        // Extract the Value cell as is - it should already be properly resolved
+        // by the InputResolver based on variable type (number, string, boolean etc.)
+        const value = row.cells.Value
+
+        // Store the correctly typed value in the result object
+        acc[row.cells.Key] = value
+      }
+      return acc
+    },
+    {} as Record<string, any>
+  )
+}
 
 export interface RequestParams {
   url: string
